+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# Unreleased
+
 # 6.0.0
 
 * Fix: Update `SentryUser` according to docs (#561)